requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
absl-py==0.11.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.0
chardet==3.0.4
cycler==0.10.0
cython==0.29.21
cython-bbox==0.1.3
fairscale
flake8==3.8.4
flake8-import-order==0.18.1
fsspec==0.8.5
future==0.18.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
imageio==2.9.0
importlib-metadata==3.3.0
iniconfig==1.1.1
joblib==1.0.1
kiwisolver==1.3.1
lap==0.4.0
llvmlite==0.36.0
markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
motmetrics==1.2.0
multidict==5.1.0
networkx==2.5.1
numba==0.53.1
oauthlib==3.1.0
omegaconf==2.0.6
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
packaging==20.8
pandas==1.2.0
pluggy==0.13.1
progress==1.5
protobuf==3.14.0
py==1.10.0
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pyquaternion==0.9.9
pytest==6.2.1
pytest-benchmark==3.2.3
python-dateutil==2.8.1
pytorch-lightning==1.1.5
pywavelets==1.1.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.18.1
scikit-learn==0.22.2
scipy==1.5.4
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
tifffile==2021.4.8
timm==0.3.4
toml==0.10.2
typing-extensions==3.7.4.3
werkzeug==1.0.1
xmltodict==0.12.0
yacs==0.1.8
yarl==1.6.3
zipp==3.4.0
